moved action to core, added support for custom name in decorator
